requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
python-dotenv==1.1.0
Werkzeug==3.1.3
openai==1.82.0
python-docx==1.1.2
markdown==3.7
gunicorn==23.0.0
email-validator==2.2.0
